compiler/requirements.txt
tensorboard==2.6
oauthlib==3.0.0
colorama==0.2.5
PyYAML==3.10
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
distro==1.7.0
flatbuffers==1.12
fonttools==4.34.4
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
networkx==2.8.4
numpy==1.23.1
opencv-python==4.7.0.68
packaging==21.3
pandas==1.4.3
pexpect==4.8.0
Pillow==7.2.0
protobuf==3.12.2
ptyprocess==0.7.0
pycocotools==2.0.4
pydot==1.4.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
scikit-build==0.15.0
scipy==1.8.1
six==1.16.0
tabulate==0.8.10
tensorboardX==2.1
typing_extensions==4.3.0
urllib3==1.26.10
onnxsim==0.4.0
adamp
wget
